Issue #14937: Migrated section 3.3.3 Ordering And Spacing to use chapter wise testing
