+++ src/it/java/com/google/checkstyle/test/chapter3filestructure/rule333orderingandspacing/OrderingAndSpacingTest.java
@@ -1,0 +1,85 @@
+///////////////////////////////////////////////////////////////////////////////////////////////
+// checkstyle: Checks Java source code and other text files for adherence to a set of rules.
+// Copyright (C) 2001-2024 the original author or authors.
+//
+// This library is free software; you can redistribute it and/or
+// modify it under the terms of the GNU Lesser General Public
+// License as published by the Free Software Foundation; either
+// version 2.1 of the License, or (at your option) any later version.
+//
+// This library is distributed in the hope that it will be useful,
+// but WITHOUT ANY WARRANTY; without even the implied warranty of
+// MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the GNU
+// Lesser General Public License for more details.
+//
+// You should have received a copy of the GNU Lesser General Public
+// License along with this library; if not, write to the Free Software
+// Foundation, Inc., 59 Temple Place, Suite 330, Boston, MA  02111-1307  USA
+///////////////////////////////////////////////////////////////////////////////////////////////
+
+package com.google.checkstyle.test.chapter3filestructure.rule333orderingandspacing;
+
+import org.junit.jupiter.api.Test;
+
+import com.google.checkstyle.test.base.AbstractGoogleModuleTestSupport;
+
+public class OrderingAndSpacingTest extends AbstractGoogleModuleTestSupport {
+
+    private static final String[] MODULE = {
+        "CustomImportOrder",
+    };
+
+    @Override
+    protected String getPackageLocation() {
+        return "com/google/checkstyle/test/chapter3filestructure/rule333orderingandspacing";
+    }
+
+    @Test
+    public void testCustomImport1() throws Exception {
+        final String filePath = getPath("InputOrderingAndSpacing1.java");
+        verifyWithConfigParser(MODULE, filePath);
+    }
+
+    @Test
+    public void testCustomImport2() throws Exception {
+        final String filePath = getPath("InputOrderingAndSpacing2.java");
+        verifyWithConfigParser(MODULE, filePath);
+    }
+
+    @Test
+    public void testCustomImport3() throws Exception {
+        final String filePath = getPath("InputOrderingAndSpacing3.java");
+        verifyWithConfigParser(MODULE, filePath);
+    }
+
+    @Test
+    public void testCustomImport4() throws Exception {
+        final String filePath = getPath("InputOrderingAndSpacing4.java");
+        verifyWithConfigParser(MODULE, filePath);
+    }
+
+    @Test
+    public void testCustomImport5() throws Exception {
+        final String filePath = getPath("InputOrderingAndSpacing5.java");
+        verifyWithConfigParser(MODULE, filePath);
+    }
+
+    @Test
+    public void testValid() throws Exception {
+        final String filePath = getPath("InputOrderingAndSpacingValid.java");
+        verifyWithConfigParser(MODULE, filePath);
+    }
+
+    @Test
+    public void testValid2() throws Exception {
+        final String filePath = getPath("InputOrderingAndSpacingValid2.java");
+        verifyWithConfigParser(MODULE, filePath);
+    }
+
+    @Test
+    public void testValidGoogleStyleOrderOfImports() throws Exception {
+        final String filePath = getPath("InputOrderingAndSpacingNoImports.java");
+        verifyWithConfigParser(MODULE, filePath);
+    }
+
+}
