--- src/it/resources/com/google/checkstyle/test/chapter3filestructure/rule333orderingandspacing/InputCustomImportOrder1.java
@@ -1,19 +1,0 @@
-package com.google.checkstyle.test.chapter3filestructure.rule333orderingandspacing;
-
-import static java.io.File.createTempFile;
-import static java.awt.Button.ABORT; //warn
-import static javax.swing.WindowConstants.*;
-import java.awt.Button; //warn
-import java.awt.Frame;
-import java.awt.Dialog; //warn
-import java.awt.event.ActionEvent;
-import javax.swing.JComponent;
-import javax.swing.JTable;
-import java.io.File; //warn
-import java.io.IOException; //warn
-import java.io.InputStream; //warn
-import java.io.Reader; //warn
-
-import com.google.common.base.Ascii; //warn
-
-public class InputCustomImportOrder1 {}
