+++ src/it/resources/com/google/checkstyle/test/chapter3filestructure/rule333orderingandspacing/InputOrderingAndSpacing4.java
@@ -1,0 +1,23 @@
+package com.google.checkstyle.test.chapter3filestructure.rule333orderingandspacing;
+
+import static java.io.File.createTempFile;
+import static java.awt.Button.ABORT;
+// violation above '.* 'java.awt.Button.ABORT' .* Should be before 'java.io.File.createTempFile'.'
+
+import static javax.swing.WindowConstants.*;
+// violation above 'Extra separation in import group before 'javax.swing.WindowConstants.*''
+
+import com.google.checkstyle.test.chapter2filebasic.rule21filename.*;
+import com.google.checkstyle.test.chapter3filestructure.rule3sourcefile.*;
+import com.google.common.reflect.*;
+import java.util.List;
+
+
+import java.util.StringTokenizer;
+// violation above 'Extra separation in import group before 'java.util.StringTokenizer''
+
+import java.util.concurrent.AbstractExecutorService;
+// violation above 'Extra separation .* before 'java.util.concurrent.AbstractExecutorService''
+
+public class InputOrderingAndSpacing4 {
+}
