+++ src/it/resources/com/google/checkstyle/test/chapter3filestructure/rule333orderingandspacing/InputOrderingAndSpacingNoImports.java
@@ -1,4 +1,4 @@
 package com.google.checkstyle.test.chapter3filestructure.rule333orderingandspacing;
 
-public class InputCustomImportOrderNoImports {
+public class InputOrderingAndSpacingNoImports {
 }
