+++ src/it/resources/com/google/checkstyle/test/chapter3filestructure/rule333orderingandspacing/InputOrderingAndSpacingValid2.java
@@ -1,14 +1,15 @@
 package com.google.checkstyle.test.chapter3filestructure.rule333orderingandspacing;
-// it is not forbidden to have extra lines (more than one) between package and import group
-
 
 import static com.puppycrawl.tools.checkstyle.utils.AnnotationUtil.containsAnnotation;
 import static com.puppycrawl.tools.checkstyle.utils.AnnotationUtil.getAnnotation;
+// comments
 
+// comments
 import com.puppycrawl.tools.checkstyle.checks.design.FinalClassCheck;
 import com.puppycrawl.tools.checkstyle.checks.design.ThrowsCountCheck;
 import com.puppycrawl.tools.checkstyle.checks.design.VisibilityModifierCheck;
 import java.util.Arrays;
+// comments
 import java.util.BitSet;
 import java.util.Map;
 import java.util.Map.Entry;
@@ -16,5 +17,5 @@ import java.util.NoSuchElementException;
 import javax.accessibility.Accessible;
 import org.apache.commons.beanutils.converters.ArrayConverter;
 
-public class InputCustomImportOrderValid {
+public class InputOrderingAndSpacingValid2 {
 }
